+++ src/mkvtoolnix-gui/main_window/preferences_dialog.h
@@ -92,6 +92,7 @@ protected:
   void setupEnableMuxingTracksByType();
   void setupEnableMuxingTracksByLanguage();
   void setupMergeAddingAppendingFilesPolicy();
+  void setupMergeWarnMissingAudioTrack();
   void setupHeaderEditorDroppedFilesPolicy();
   void setupJobsRunPrograms();
   void setupFont();